THIS FILE CONTAINS PART OF MY R-SCRIPT FOR MBS1005
I excluded chunks of code which I don't think are that useful for us.
(e.g. I removed the biggest part of the PCA plots, because I didn't have enough time in the end to make a loop, so it was just tediously long).

Actually,  the only thing that might have some added value is the code for the volcano plots in part 3.

(Also: I have no idea if what I did for part 5 is even remotely right.)