Included an alternative file for the tan module
Included an alternative file for the tan module: derived from the PPI network with the default confindence cutoff of 0.4. We can compare this with the other excel file for the tan module (where the cutoff was 0.7), see if the same genes of interest come up and/or if the lower cutoff returns more interesting results to look at.