Tan major component network images
PNG and PDF file for tan PPI network. Legend is included for betweenness centrality. Information on the color, size, and black border can be added in the firgure description, but pasting more infomation about the border in the figure itself is an option when there is no more important stuff left to do (also it was not possible to change 'Node Fill Color: Betweenness Centraility' text in Cytoscape, but I'll see if I can paste over that later).